Cleanup multiple host test
